Move most Router logic into WasmKeeper
