+++ packages/multi-test/src/lib.rs
@@ -5,7 +5,7 @@ mod test_helpers;
 mod transactions;
 mod wasm;
 
-pub use crate::app::{next_block, parse_contract_addr, App, Router};
+pub use crate::app::{next_block, App, Router};
 pub use crate::bank::{Bank, SimpleBank};
 pub use crate::contracts::{Contract, ContractWrapper};
-pub use crate::wasm::WasmKeeper;
+pub use crate::wasm::{parse_contract_addr, Wasm, WasmKeeper};
